requirements.txt
nfcpy
soco
# uncomment for Chromecast support
#pychromecast
pyudev
